% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-tests.R
\name{expect_pdf_matches_cdf}
\alias{expect_pdf_matches_cdf}
\title{Verify a pdf matches a cdf.}
\usage{
expect_pdf_matches_cdf(pdf, cdf, domain, arg_bounds, n_arg_points)
}
\arguments{
\item{pdf}{the putative probability density function.}

\item{cdf}{the putative cumulative density function.}

\item{domain}{c(lower, upper) indicating the pdf's domain.}

\item{arg_bounds}{a list with one named element per argument to f,
containing a 2-vector with the lower and upper bound for
parameter values of f to consider.}

\item{n_arg_points}{An integer indicating how many points to generate
within the parameter bounds.}
}
\description{
Verify a pdf matches a cdf.
}

