% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-stan.R
\name{find_stan_functions}
\alias{find_stan_functions}
\title{Function for reading in the list of Stan function files.  These are
not proper Stan programs as they are missing the 'functions' wrapper
as well as the model block.}
\usage{
find_stan_functions()
}
\value{
vector of strings with paths of Stan function files in the
package.
}
\description{
Function for reading in the list of Stan function files.  These are
not proper Stan programs as they are missing the 'functions' wrapper
as well as the model block.
}

