% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement-gamma.R
\name{gamma_pdf_1}
\alias{gamma_cdf_1}
\alias{gamma_lcdf_1}
\alias{gamma_lpdf_1}
\alias{gamma_pdf_1}
\title{Naive implementation of the Gamma density.}
\usage{
gamma_pdf_1(x, alpha, beta)

gamma_lpdf_1(x, alpha, beta)

gamma_cdf_1(x, alpha, beta)

gamma_lcdf_1(x, alpha, beta)
}
\arguments{
\item{x}{Value to evaluate density at.}

\item{alpha}{Shape parameter.}

\item{beta}{Scale parameter.}
}
\description{
Naive implementation of the Gamma density.
}
\section{Functions}{
\itemize{
\item \code{gamma_lpdf_1}: Naive implementation of the Gamma log
density.

\item \code{gamma_cdf_1}: Naive implementation of the Gamma
cumulative density.

\item \code{gamma_lcdf_1}: Naive implementation of the Gamma log
cumulative density.
}}

