% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement-generalized-gamma.R
\name{generalized_gamma_pdf_1}
\alias{generalized_gamma_cdf_1}
\alias{generalized_gamma_lcdf_1}
\alias{generalized_gamma_lpdf_1}
\alias{generalized_gamma_pdf_1}
\title{Naive implementation of the generalized Gamma density.}
\usage{
generalized_gamma_pdf_1(x, alpha, beta, nu)

generalized_gamma_lpdf_1(x, alpha, beta, nu)

generalized_gamma_cdf_1(x, alpha, beta, nu)

generalized_gamma_lcdf_1(x, alpha, beta, nu)
}
\arguments{
\item{x}{Value to evaluate density at.}

\item{alpha}{Shape parameter.}

\item{beta}{Scale parameter.}

\item{nu}{Tail parameter.}
}
\description{
Naive implementation of the generalized Gamma density.
}
\section{Functions}{
\itemize{
\item \code{generalized_gamma_lpdf_1}: Naive implementation of the
generalized Gamma log density.

\item \code{generalized_gamma_cdf_1}: Naive implementation of the
generalized Gamma cumulative density.

\item \code{generalized_gamma_lcdf_1}: Naive implementation of the
generalized Gamma log cumulative density.
}}

