% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement-generalized-gamma.R
\name{generalized_gamma_pdf_3}
\alias{generalized_gamma_cdf_3}
\alias{generalized_gamma_lcdf_3}
\alias{generalized_gamma_lpdf_3}
\alias{generalized_gamma_pdf_3}
\title{Naive implementation of the generalized Gamma density from flexsurv
package (Stacy 1962)}
\usage{
generalized_gamma_pdf_3(x, alpha, beta, nu)

generalized_gamma_lpdf_3(x, alpha, beta, nu)

generalized_gamma_cdf_3(x, alpha, beta, nu)

generalized_gamma_lcdf_3(x, alpha, beta, nu)
}
\arguments{
\item{x}{Value to evaluate density at.}

\item{alpha}{Shape parameter.}

\item{beta}{Scale parameter.}

\item{nu}{Tail parameter.}
}
\description{
Naive implementation of the generalized Gamma density from flexsurv
package (Stacy 1962)
}
\section{Functions}{
\itemize{
\item \code{generalized_gamma_lpdf_3}: Naive implementation of the
generalized Gamma log density from flexsurv package (Stacy 1962)

\item \code{generalized_gamma_cdf_3}: Naive implementation of the
generalized Gamma cumulative density from flexsurv package (Stacy 1962)

\item \code{generalized_gamma_lcdf_3}: Naive implementation of the
generalized Gamma log cumulative density from flexsurv package (Stacy 1962)
}}

