% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement-generalized-gamma.R
\name{generalized_gamma_pdf_4}
\alias{generalized_gamma_cdf_4}
\alias{generalized_gamma_lcdf_4}
\alias{generalized_gamma_lpdf_4}
\alias{generalized_gamma_pdf_4}
\title{Implementation of the generalized Gamma density from R flexsurv
package}
\usage{
generalized_gamma_pdf_4(x, k, mu, sigma)

generalized_gamma_lpdf_4(x, k, mu, sigma)

generalized_gamma_cdf_4(x, k, mu, sigma)

generalized_gamma_lcdf_4(x, k, mu, sigma)
}
\arguments{
\item{x}{Value to evaluate density at.}

\item{k}{Shape parameter.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}
}
\description{
Implementation of the generalized Gamma density from R flexsurv
package
}
\section{Functions}{
\itemize{
\item \code{generalized_gamma_lpdf_4}: Implementation of the generalized
Gamma log density from R flexsurv package.

\item \code{generalized_gamma_cdf_4}: Implementation of the generalized
Gamma cumulative density from R flexsurv package.

\item \code{generalized_gamma_lcdf_4}: Implementation of the generalized
Gamma cumulative log density from R flexsurv package.
}}

