% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement-log-gamma.R
\name{generalized_log_gamma_pdf_1}
\alias{generalized_log_gamma_pdf_1}
\title{Density of the log generalized gamma directly from Lawless (1992)}
\usage{
generalized_log_gamma_pdf_1(x, k, mu, sigma)
}
\arguments{
\item{x}{Value to evaluate density at.}

\item{k}{Shape parameter.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}
}
\description{
Density of the log generalized gamma directly from Lawless (1992)
}

