% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-stan.R
\name{generate_declarations}
\alias{generate_declarations}
\title{Function for pulling out declarations from Stan function files.}
\usage{
generate_declarations(s)
}
\description{
Function for pulling out declarations from Stan function files.
}

