% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-tests.R
\name{generate_parameter_points}
\alias{generate_parameter_points}
\title{Create a set of points in parameter space.}
\usage{
generate_parameter_points(arg_bounds, n_points)
}
\arguments{
\item{arg_bounds, }{n_arg_points.}
}
\value{
data frame of points uniformly distributed in parameter
        space.
}
\description{
Create a set of points in parameter space.
}

