% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-numerical.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{log_sum_exp implementation as suggested by Bill Dunlap
and modified to deal with Inf.}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{A vector of numbers to exp-sum-log}
}
\value{
log-sum-exp of the x argument.
}
\description{
log_sum_exp implementation as suggested by Bill Dunlap
and modified to deal with Inf.
}

