% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-stan.R
\name{merge_stan_functions}
\alias{merge_stan_functions}
\title{Function for generating a proper dummy Stan functions file with
functions block from function files available in the installed
package.}
\usage{
merge_stan_functions()
}
\description{
Function for generating a proper dummy Stan functions file with
functions block from function files available in the installed
package.
}

