% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{weibull_lpdf}
\alias{gamma_lpdf}
\alias{generalized_gamma_lpdf}
\alias{weibull_lpdf}
\alias{weibull_uniform_mixture_lpdf}
\title{API function for weibull log density.}
\usage{
weibull_lpdf(x, ..., method = "naive")

gamma_lpdf(x, ..., method = "naive")

weibull_uniform_mixture_lpdf(x, ..., method = "naive")

generalized_gamma_lpdf(x, ..., method = "Lawless-1992")
}
\arguments{
\item{x}{Numeric value to evaluate the log density at.}

\item{...}{Parameterization-specific parameter parameter values.}

\item{method}{Length one character vector for name of
parameterization.}
}
\value{
Weibull log density at x.
}
\description{
API function for weibull log density.
}
\section{Functions}{
\itemize{
\item \code{gamma_lpdf}: API function for gamma log density.

\item \code{weibull_uniform_mixture_lpdf}: API function for Weibull-uniform mixture log density.

\item \code{generalized_gamma_lpdf}: API function for generalized gamma log density.
}}

