% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{weibull_pdf}
\alias{gamma_pdf}
\alias{generalized_gamma_pdf}
\alias{weibull_pdf}
\alias{weibull_uniform_mixture_pdf}
\title{API function for weibull density.}
\usage{
weibull_pdf(x, ..., method = "naive")

gamma_pdf(x, ..., method = "naive")

weibull_uniform_mixture_pdf(x, ..., method = "naive")

generalized_gamma_pdf(x, ..., method = "Lawless-1992")
}
\arguments{
\item{x}{Numeric value to evaluate the density at.}

\item{...}{Parameterization-specific parameter parameter values.}

\item{method}{Length one character vector for name of
parameterization.}
}
\value{
Weibull density at x.
}
\description{
API function for weibull density.
}
\section{Functions}{
\itemize{
\item \code{gamma_pdf}: API function for gamma density.

\item \code{weibull_uniform_mixture_pdf}: API function for Weibull-uniform mixture density.

\item \code{generalized_gamma_pdf}: API function for generalized gamma density.
}}

