% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement-weibull.R
\name{weibull_pdf_1}
\alias{weibull_cdf_1}
\alias{weibull_lcdf_1}
\alias{weibull_lpdf_1}
\alias{weibull_pdf_1}
\title{Naive implementation of the Weibull density.}
\usage{
weibull_pdf_1(x, alpha, beta)

weibull_lpdf_1(x, alpha, beta)

weibull_cdf_1(x, alpha, beta)

weibull_lcdf_1(x, alpha, beta)
}
\arguments{
\item{x}{Value to evaluate density at.}

\item{alpha}{Shape parameter.}

\item{beta}{Scale parameter.}
}
\description{
Naive implementation of the Weibull density.
}
\section{Functions}{
\itemize{
\item \code{weibull_lpdf_1}: Naive implementation of the Weibull log-density.

\item \code{weibull_cdf_1}: Naive implementation of Weibull cumulative density.

\item \code{weibull_lcdf_1}: Naive implementation of Weibull log cumulative density.
}}

