% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement-weibull-uniform-mixture.R
\name{weibull_uniform_mixture_pdf_1}
\alias{weibull_uniform_mixture_cdf_1}
\alias{weibull_uniform_mixture_lcdf_1}
\alias{weibull_uniform_mixture_lpdf_1}
\alias{weibull_uniform_mixture_pdf_1}
\title{Naive implementation of the Weibull-uniform mixture density.}
\usage{
weibull_uniform_mixture_pdf_1(x, max, q, alpha, beta)

weibull_uniform_mixture_lpdf_1(x, max, q, alpha, beta)

weibull_uniform_mixture_cdf_1(x, max, q, alpha, beta)

weibull_uniform_mixture_lcdf_1(x, max, q, alpha, beta)
}
\arguments{
\item{x}{Value to evaluate density at.}

\item{max}{Maximum value x could take.}

\item{q}{Mixing parameter.}

\item{alpha}{Shape parameter.}

\item{beta}{Scale parameter.}
}
\description{
Naive implementation of the Weibull-uniform mixture density.
}
\section{Functions}{
\itemize{
\item \code{weibull_uniform_mixture_lpdf_1}: Naive implementation of the Weibull-uniform mixture log density.

\item \code{weibull_uniform_mixture_cdf_1}: Naive implementation of the
Weibull-uniform mixture cumulative density.

\item \code{weibull_uniform_mixture_lcdf_1}: Naive implementation of the
Weibull-uniform mixture log cumulative density.
}}

